+++ src/api/address.js
@@ -1,47 +1,26 @@
 import request from '@/utils/request'
-import store from '@/store'
 // 获取省市区
 export const getRegionProvince = () => {
-  return request.get('/my/region/province', {
-    headers: {
-      Authorization: store.getters.getToken
-    }
-  })
+  return request.get('/my/region/province')
 }
 
 export const getRegionCity = (regionId) => {
-  return request.get(`/my/region/city/${regionId}`, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
-  })
+  return request.get(`/my/region/city/${regionId}`)
 }
 
 export const getRegionCounty = (regionId) => {
-  return request.get(`/my/region/county/${regionId}`, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
-  })
+  return request.get(`/my/region/county/${regionId}`)
 }
 
 // 1.获取收获地址列表
 export const getAddressList = () => {
-  return request.get('/my/address/list', {
-    headers: {
-      Authorization: store.getters.getToken
-    }
-  })
+  return request.get('/my/address/list')
 }
 
 // 2.添加收获地址
 export const setAddAddress = (form) => {
   return request.post('/my/add/address', {
     form
-  }, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
   })
 }
 
@@ -49,10 +28,6 @@ export const setAddAddress = (form) => {
 export const clearAddress = (addressId) => {
   return request.post('/my/clear/address', {
     addressId
-  }, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
   })
 }
 
@@ -61,10 +36,6 @@ export const updateAddress = (addressId, form) => {
   return request.post('/my/update/address', {
     addressId,
     form
-  }, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
   })
 }
 
@@ -72,9 +43,5 @@ export const updateAddress = (addressId, form) => {
 export const setDefaultAddress = (addressId) => {
   return request.post('/my/default/address', {
     addressId
-  }, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
   })
 }