+++ src/api/cart.js
@@ -1,4 +1,3 @@
-import store from '@/store'
 import request from '@/utils/request'
 // 1. 添加到购物车
 export const setAddCart = (goodsId, specValueIds, quantity) => {
@@ -6,19 +5,11 @@ export const setAddCart = (goodsId, specValueIds, quantity) => {
     goodsId,
     specValueIds,
     quantity
-  }, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
   })
 }
 // 获取购物车列表数据
 export const getCartList = () => {
-  return request.get('/my/cart/list', {
-    headers: {
-      Authorization: store.getters.getToken
-    }
-  })
+  return request.get('/my/cart/list')
 }
 
 // 更新购物车商品数量
@@ -26,10 +17,6 @@ export const setUpdateCount = (cartId, quantity) => {
   return request.post('/my/cart/update', {
     cartId,
     quantity
-  }, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
   })
 }
 
@@ -37,9 +24,5 @@ export const setUpdateCount = (cartId, quantity) => {
 export const deleteCartList = (cartIds) => {
   return request.post('/my/cart/clear', {
     cartIds
-  }, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
   })
 }