+++ src/api/pay.js
@@ -1,5 +1,4 @@
 import request from '@/utils/request'
-import store from '@/store'
 // 获取订单结算数据
 
 export const getChekoutOrderData = (mode, obj) => {
@@ -7,20 +6,13 @@ export const getChekoutOrderData = (mode, obj) => {
     params: {
       mode,
       ...obj
-    },
-    headers: {
-      Authorization: store.getters.getToken
     }
   })
 }
 
 // 获取支付类型与余额
 export const getPayModeData = () => {
-  return request.get('/my/pay/mode', {
-    headers: {
-      Authorization: store.getters.getToken
-    }
-  })
+  return request.get('/my/pay/mode')
 }
 
 // 提交订单
@@ -28,9 +20,5 @@ export const submitOrder = (mode, obj) => {
   return request.post('/my/checkout/submit', {
     mode,
     ...obj
-  }, {
-    headers: {
-      Authorization: store.getters.getToken
-    }
   })
 }