+++ src/components/BottomPopup.vue
@@ -40,7 +40,9 @@
 <script>
 import { setAddCart } from '@/api/cart'
 import SpecSelector from '@/components/SpecSelector.vue'
+import loginConfirm from '@/mixins/loginConfirm.js'
 export default {
+  mixins: [loginConfirm],
   components: {
     SpecSelector
   },
@@ -70,31 +72,6 @@ export default {
     close () {
       this.$emit('input', false)
     },
-    handleConfirm () {
-      if (!this.$store.getters.getToken) {
-        // 在Vue组件中
-        this.$modal.confirm({
-          title: '温馨提示',
-          message: '你还没有登录，无法进行此操作哦！',
-          confirmButtonColor: '#ee0a24',
-          cancelBtnText: '去逛逛',
-          confirmBtnText: '去登录',
-          closerOnClickModal: false
-        })
-          .then(() => {
-            this.$router.replace({
-              path: '/login',
-              query: {
-                url: this.$route.fullPath // 当前的路径地址
-              }
-            })
-            this.$emit('input', false) // 关闭底层弹框
-          })
-          .catch(() => { })
-        return true
-      }
-      return false
-    },
     async addCart () {
       try {
         if (this.handleConfirm()) return
@@ -108,8 +85,8 @@ export default {
       }
     },
     buyNow () {
-      // 判断是否又token权证
-
+      // 判断是否有token权证
+      if (this.handleConfirm()) return
       this.$router.push({
         path: '/pay',
         query: {